Refactor exact search enabled test to test equality not flag
